Allow players to specify which playlist to pull tracks from
This commit changes the constructors of LyricGuess and SongGuess
to take a nullable Playlist object. This is then used to randomly
select tracks from that specific playlist, if non-null, or default
to randomly sampling from all of the user's playlist.

Additionally, this commit adds a stop command to all interactive
games. This allows users to stop the game at any point instead of
having to wait for its completion.